+++ models/UserModel.php
@@ -49,6 +49,9 @@ class UserModel extends BaseModel
     public function deleteUserById($id)
     {
         $id = $this->decryptID($id);
+        if($id==null){
+            return false;
+        }
         $sql = 'DELETE FROM users WHERE id = ' . $id;
         return $this->delete($sql);
     }
